Moved enums package to definitions package, Reimplemented ChatRoom entity
