+++ src/main/java/pl/abbl/reactchat/definitions/enums/ChatRoomStatus.java
@@ -1,0 +1,7 @@
+package pl.abbl.reactchat.definitions.enums;
+
+public enum ChatRoomStatus {
+    OPEN,
+    CLOSED,
+    REMOVED
+}
